Restoring WebView's context to the Activity instead of the app context
Using the application context for a WebView will result in the webview failing to handle dialog pickers (like an HTML SELECT tag for instance).

Also, removed the import for `View.FOCUSABLE` and lint checks can't determine the API level for the import, but it does when it is used inside of the `if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)` check. (one less yellow warning in the IDE)