Remove destruction migration fallback (#193)
We can no longer rely on blowing away the DB when the schema changes. We need to properly write DB migrations from now on, and I'd like to be made aware of this need as early as possible.

Leaving the fallback in place is dangerous, as we are going to end up changing the DB, releasing and blowing away all the user's data.