Consolidate `HomeActivity` and `BrowserActivity` landing pages
This removes HomeActivity entirely and allows the user to immediately start typing their search when launched.

One consequence of this is that we have to work a little harder to destroy the webview since it won't be destroyed naturally as before, by destroying the browser activity. As such, we basically reset the view state of browser activity instead. A change here is that now the WebView can be in a destroyed state which is dangerous; any interations with it after this would result in a crash. So I've made the WebView nullable to have the tools better help protect against this state.