Feature/consolidated landing page (#208)
* Update activity browser layout to have DDG logo as "background"

* Add adjustResize to browser activity

* Consolidate `HomeActivity` and `BrowserActivity` landing pages

This removes HomeActivity entirely and allows the user to immediately start typing their search when launched.

One consequence of this is that we have to work a little harder to destroy the webview since it won't be destroyed naturally as before, by destroying the browser activity. As such, we basically reset the view state of browser activity instead. A change here is that now the WebView can be in a destroyed state which is dangerous; any interations with it after this would result in a crash. So I've made the WebView nullable to have the tools better help protect against this state.

* Formatting only

* Update documentation on class now that its responsibility has changed

* Formatting

* Fix package name for UserAgentProvider

* Restore handling of external URLs

* Fix problem with toolbar being accessed before it was created

This was causing the fire button not to be shown when it was expected to be visible

* Remove redundant variable in view state constructor; tidy imports

* Delay keyboard showing a little so it loads with the activity

* Format BrowserActivity manifest definition

* Remove some of the "replace existing search" logic where possible

* Break apart intent interrogation for resetting state vs consuming query

* Use @color/white instead of #ffffff