Update libraries - required to fix a crash on Android lollipop
Didn't want to go updating libraries as part of this, but noticed the crash when trying out the code on a Lollipop emulator.

Upgrading to 27.1.0 of the support library isn't enough - we also need to bring in an update to architecture components extensions

This now introduces a very sensible change to the `onCreateViewHolder` API which is that the parent is now delivered marked as non-null. While awesome, it requires our methods to be marked with the `?` to compile.