Correct stored ATB if it still has the old app's format
Previously, MA was hardcoded on the end of the ATB param. We no longer want this functionality, so we'll need to check for previously stored ATB values explicitly and remove the suffix if it exists. In this scenario, we'll also want to set the stored variant to the default variant to avoid the user being reallocated to a new variant.