Feature/default browser (#268)
* Integrate new onboarding screen for default browser (WIP)

The color transition between screens still isn't perfect; it works unless you switch directions mid-swipe and then it goes a bit awry. Will fix soon.

* Add default browser switch into settings

* Fix tests

* Add default browser detector and wire into settings and onboarding

* Stop untoggling the default browser toggle; UX works better to allow it

It will correct itself if necessary when the user returns from the settings screen.

* Add default browser banner, timed to show after certain delay

* Change to 8dp corner radius on default app info activity

* Wire in default browser info activity

* Restore missing string removed in a refactor

* Remove developer hacks to show notification banner

* Fix color transition on onboarding pages

* Make onboarding status bar transparent to avoid updating its color

* Tidy up logic whether to show onboarding / home screens

Home is always shown, so we might as well remove the cycle of starting onboarding, waiting for it to finish, checking the request code in onActivityResult, calling back to viewModel, then waiting for a new command to be observed instructing us to launch home.

* Code formatting

* Remove redundant class

* Remove suspected flaky test

* Only show default browser UI when variant supports it

* Remove noisy log statement

* Code formatting

* Add some tests for the defined variants

* Code formatting as per PR feedback observations

* More travis experimentation

* Store ATB separately from variant

* Revert "Remove suspected flaky test"

This reverts commit 1c4fd6076b99e8a115e45183353efac8d2c4cf0a.

* Store ATB separately from variant and fix code formatting

* Correct stored ATB if it still has the old app's format

Previously, MA was hardcoded on the end of the ATB param. We no longer want this functionality, so we'll need to check for previously stored ATB values explicitly and remove the suffix if it exists. In this scenario, we'll also want to set the stored variant to the default variant to avoid the user being reallocated to a new variant.

* Increase touch area for the notification banner's dismiss button

* Change size of close button on default browser info activity

* Code formatting