Add canGoBack() and canGoForward() into view state so it is accounted for (#302)
A regression meant that the view state was checked to see if it was identical to the last seen view state. If it was, rendering would skip. But for these forward/back buttons, they also relied on WebView state; not just the view state. So we were erroneously skipping updates for them because the view state was identical.

The fix is to include these in the view state itself, meaning it won't skip the rendering if canGoBack() or canGoForward() has changed.

onProgressChanged is the earliest place where the webView starts reporting the correct new states. But I've included the update in loadingFinished() too just in case onProgressChanged could be bypassed somehow.