Add custom url scheme handling
We also show a dialog of some sorts before launching the external app so that the user can cancel if they didn't mean to launch that external app.

If there is only one app that can be launched for that link, we show an yes/no AlertDialog. Otherwise, we let the system Activity picker kick in. Either way, the user could cancel if they wanted to.