Feature/custom url schemes (#322)
* Add custom url scheme handling

We also show a dialog of some sorts before launching the external app so that the user can cancel if they didn't mean to launch that external app.

If there is only one app that can be launched for that link, we show an yes/no AlertDialog. Otherwise, we let the system Activity picker kick in. Either way, the user could cancel if they wanted to.

* Handle custom scheme entered directly into url bar

* Add error message to strings file

* Code tidy

* Handle intent fallback URLs if a compatible app isn't installed

* Adding special handling for "about:blank" link types

Otherwise it's possible to create an infinite loop on the likes of m.yelp.com when trying to launch the native app.