Use onPageCommitVisible as webview hook to update URL bar
This is to remove the ability for a malicious website to perform a type of phishing attack, whereby they would show content in the webview but attempt to load a page they know will take a while to timeout. The result of the attack is that the URL could display a genuine looking domain (https://www.gmail.com:8080 for example) along with a genuine looking Google-login page but that would send the login details to the malicious site.