Make use of onProgressChanged instead on onPageCommitVisible
We erroneously relied on onPageCommitVisible() but it was only ever called on API 23 or up, breaking functionality for Lollipop users.

We put in a patch, which made use of onPageCommitVisible() for newer users only. Now, we're stopping relying on onPageCommitVisible altogether and making use of onProgressChanged instead for the functionality.