Move from IO scheduler to single-threaded executor for DB work
This is to address the situation where we might have concurrent requests to open a new tab (one the blank, default tab and the other the one the user really wants to see). With the IO scheduler, there is the possibility of adding both, and then selecting the default tab as the one to show the user, resulting in the desired tab being relegated to the background.

Moving to a single-threaded executor should let us better control the DB-related actions like adding and selecting a tab. There are likely even better ways to do this in future, but hopefully this will suffice for now.