Change web content to be hidden when ViewState is first initialized
This should always have been the case. Leaving it init to showing the web view content can cause timing issues whereby an incoming intent could be processed and consumed too early, and thus leaves the intent liable to get "lost"; if the incoming intent is to load a URL and that load begins, but then shortly afterwards the data clearing kills all tabs, the URL the user wanted will not be shown to them.