Feature/automatic data clearing home screen shortcut fix (#413)
* Move from IO scheduler to single-threaded executor for DB work

This is to address the situation where we might have concurrent requests to open a new tab (one the blank, default tab and the other the one the user really wants to see). With the IO scheduler, there is the possibility of adding both, and then selecting the default tab as the one to show the user, resulting in the desired tab being relegated to the background.

Moving to a single-threaded executor should let us better control the DB-related actions like adding and selecting a tab. There are likely even better ways to do this in future, but hopefully this will suffice for now.

* Add RxJavaPlugin hook for single scheduler

* Change web content to be hidden when ViewState is first initialized

This should always have been the case. Leaving it init to showing the web view content can cause timing issues whereby an incoming intent could be processed and consumed too early, and thus leaves the intent liable to get "lost"; if the incoming intent is to load a URL and that load begins, but then shortly afterwards the data clearing kills all tabs, the URL the user wanted will not be shown to them.