Feature/restore partial network caching (#515)
* Move OkHttp client cache to a named directory under "cache" dir

This will effectively result in a one-time cache invalidation. But will allow us to delete the entire cache directory while preserving our (known to be non-sensitive) HTTP networking cache.

* Rename "pixel" networking dependency to be "nonCaching"

Since we want to share this across anything else which wants to make non-caching requests, we need to generify it a little so it isn't just dealing with pixels.