Generate webview image when leaving BrowserActivity
This fixes an issue whereby a page might be loaded in a tab (TAB A) but the callback for page loaded is fired before anything is visible, producing a blank WebView preview. This would normally be fine if the user hit the tab switcher button. But if they instead leave the app, then they might return to the app later launching another tab (TAB B). From TAB B, they could launch the tab switcher. TAB B preview would be re-generated, but TAB A would remain white.

This commit ensure the preview is generated when the user is leaving the screen altogether, and this removes the need to generate a bitmap every time the page loads.