Generate WebView preview when the tab fragment is hidden
This takes care of an edge case whereby a new tab is opened. In this scenario, we wouldn't trigger the webview preview generation through the whole app backgrounding, but the tab would still be switching. If the app died while viewing the other tab, the original tab would never have its preview generated.