Generate webview image when leaving BrowserActivity (#597)
* Generate webview image when leaving BrowserActivity

This fixes an issue whereby a page might be loaded in a tab (TAB A) but the callback for page loaded is fired before anything is visible, producing a blank WebView preview. This would normally be fine if the user hit the tab switcher button. But if they instead leave the app, then they might return to the app later launching another tab (TAB B). From TAB B, they could launch the tab switcher. TAB B preview would be re-generated, but TAB A would remain white.

This commit ensure the preview is generated when the user is leaving the screen altogether, and this removes the need to generate a bitmap every time the page loads.

* Generate WebView preview when the tab fragment is hidden

This takes care of an edge case whereby a new tab is opened. In this scenario, we wouldn't trigger the webview preview generation through the whole app backgrounding, but the tab would still be switching. If the app died while viewing the other tab, the original tab would never have its preview generated.

* Reduce noise of log statement

* Ensure WebView preview generated for new tab launches (better impl)

Before, we tried to utilise onHiddenChanged callback for this, but it introduced more problems than it solved. When returning to the app to launch a new tab for instance, it invoked this callback on the tab which was disappearing and sometimes then overwrote a perfectly good preview with a now blank one (if the WebView was non-null but not rendering content yet)

* Code format

* Generate WebView preview when new tab selected from the popup menu

* Add tests ensuring WebView previews are generated when expected