Update Fastlane script to contain a `release` lane.
You can use this to better automate the release process.

## Pre-Requisites
  - You need `git flow` tools installed (this would be an easy tidy-up to remove this dependency later)
  - You have `develop` checked out
  - You have a clean working directory

## Using
Start with `fastlane release` and it will prompt for new version number and release notes. It will perform the required updates to the `version.properties` and `release-notes` files, on a release branch, and will confirm everything looks right before it creates a tag and does the required merging back into `develop` and `master`.