For the global exception catcher, use `runBlocking` to preserve behavior
With the use of GlobalScope.launch, we'd potentially be some other code to run after this exception had been thrown, and before the uncaughtException function here had fully executed. Use of `runBlocking` to ensure execution waits for this function to complete makes more sense and it is likely to result in less unpredictable issues.