Use SystemClock.elapsedRealtime() for data clearing
This offers more resiliency vs System.currentTimeMillis() against someone meddling with the device clock in an effort to delay data being cleared when expected.

elapsedRealtime will starting again when the system is restarted, but in that scenario, we'd wipe the data anyway since it's a new process launch, so we don't have to worry about the recorded timestamp being smaller than the current one.