Add global RxJava error handler to look for UndeliverableExceptions (#787)
After updating the app dependencies to newest versions of all libraries, we noticed an increase in UndeliverablExceptions. This, it seems, is most likely caused by our sync job being interrupted, so we in turn dispose of the RxJava subscription. However in cases where the networking capabilities have just dropped (a potential cause of the sync job being stopped) we would then get a bunch of IO-related exceptions being called and since we'd already disposed, RxJava would have nowhere to deliver the exception to, hence it would throw an UndeliverableException.

This UndeliverableException wouldn't be caught by the RxJava global error handler (as we don't set one) so it would default to the uncaught exception handler for the thread.

This change catches and logs UndeliverableExceptions but doesn't let the app crash because of them. For all other exceptions, they continue to defer to the uncaught exception handler as before.