Adds a "Developer Settings" placeholder page to "internal" builds
For now, the dev settings is empty, but crucially only visible to internal builds. For this to work, adds gradle flavors for audience, which can be either "internal" or "external". The idea being that we can make internal builds which can offer slightly different functionality to the external builds.

Internal builds could be distributed internally only, and therefore we could do things in there that we couldn't add externally. This might include:
- performance metrics
- feature toggles, especially for more experimental things
- see additional data, stats that we don't expose in the main app UI