Add domain extraction for plaintext and encrypted requests
Note, looks like we sometimes get packets for processing from the TUN before they have enough data in them. ie, because they are being streamed, we have enough information to stream the request to the server, but not enough to evaluate locally for everything we need to extract domain name. This will need sorted.