Fix favicon display view recycling
Favicons are displayed in a number of places, eg. tab switcher, bookmarks, fireproof websites etc.
All those places use a listview adapter. The bug appears when the view holder is recycled and the new item
to be displayed in it does not have a favicon. In which case, whatever icon was set in the recycled holder is showed.

This PR:

* modifies the `FaviconManager` so that it preloads the default favicon before trying to load the new one

* Adds unit tests for it

Fixes #975

NOTE: I also considered doing this in the `FaviconDownloader` by allowing a nullable `File` in the `loadFaviconToView` method. However this is too late as results in a UX that shows the previous (holder) icon and then it changes to the proper one.