Improve the build times for the VPN project (#41)
* Improve the build times for the VPN project

The build time for incremental builds increased by 4x (from 10s -> 40s) when we added Anvil. This is due to several things:
* there is a bug in kotlin whereby [incremental compilation breaks compilre plugins](https://youtrack.jetbrains.com/issue/KT-38570). Anvil is a compiler plugin and works around this issue by just disabling incremental compilation for some of the gradle tasks, that in turn increases incremental builds

* the bug above shows up when mixing Java and Kotlin in the same module where anvil is installed. As we use annotation processors for Daggeer and Room, our vpn module hits this but

This PR improves the build time for incremental builds by a factor of 2, leaving the impact of anvil in build times to be "just" 2x for now (from 10s -> 20s) which leads to a better trade off to continue using Anvil.

As Anvil can also generate all the factories for the `@Inject` constructors, `@Inject` fields and `@Provides` methods, it is possible to split the `:vpn` module in 3 such that it both makes sense and Anvil does not need to live in a module with mix-sources.

We have split the `:vpn` module in 3:
* `:vpn-main` -> this module is very tiny and only contains the definition of the dagger components needed in the VPN feature. The `:app` will only depend on this module. The dagger annotation processor will take care of generating the code for the DI components

* `:vpn-store` -> this module will contain the anything related to the VPN room database. Room annotation processor will take care of code generation

* `:vpn` -> stays as is minus what's moved to the modules above. The main point is that we have removed the need for both Dagger and Room annotation processors here and so build times got better.

A bonus of this approach is that, now Dagger and Room annotation processors being in different modules, we also get a smaller build improve from the fact that when one generates code, saves the pass the other needs to perform to see whether it's interested in any of the new code to do further code generation.

Nuance: The `:app` also depends on the `:vpn-store`. This is a quick convenience fix for now. The reason is that the `:app` module knows about the `VpnDatabase` so that the `DaggerWorkerFactory` can create the `VpnStatsReportingWorker`. There is an easy way to improve this intreface making it more loosely coupled, but it'd require to also move the `statistics` package (for pixels) to its own library. Althogh this makes complete sense, it hasn't been done under this PR.

* Fix tests that broke when merging develop branch

* Undo removal of prepopulateTrackerCompanies