Replace use of Handlers to delay work to avoid ANR possibility (#44)
I saw an ANR happening because the main thread was blocked awaiting packet handling. Upon investigating, the cause of this was using Handler to delay the execution of closing the connection (as that will run on the main thread). Instead, switched that small part to using coroutines and delay() instead.

Co-authored-by: Craig Russell <CDRussell@users.noreply.github.com>