Use alternative method of obtaining originating app package (#1551)
packageManager.getNameForUid(uid) was mostly good, but in some cases returning a package name that wouldn't match when we queried for the associated installed app (examples include Noom and MX Player)

switching to packageManager.getPackagesForUid does a better job at getting the package right.

one nuance: the API returns a list of packages, instead of just one. in most cases, the list should have one entry. but it's possible there will be more there for some entries, so we always take the first. (order is unclear, and I haven't seen any examples of more than 1 entry appearing so we'll need to see once examples show up if we should take the last, or need a different approach altogether)