Catch IllegalArgumentException in deviceToNetworkProcessing (#1622)
Task/Issue URL: https://app.asana.com/0/1199371158290272/1201501533153172/f

### Description
We're seeing a crash in the UDP pipeline, in the `deviceToNetworkProcessing` method when trying to write in the datagram channel.

The exception is `IllegalArgumentException` but we're only catching `IOExceptions`.

At the time of writing this is affecting 284 users, which is a lot when considering how many we've got in the beta waitlist.

### Steps to test this PR
We can't reproduce this issue, so just code review.