Add reported app default protection info to breakage report (#1672)
Task/Issue URL: https://app.asana.com/0/1199371158290272/1201539333793164/f

### Description
During the breakage it is benefitial to know whether the reported app was enabled by default of not.

This PR adds a field inside the `vpnTrackerLists` JSON object that contains whether the package ID was unprotected (ie. inside the exclusion list) or not. 
The new field is called `reportedAppUnprotectedByDefault` and can take a "true" or "false" value.

### Steps to test this PR

_Test_
- [x] install from this branch
- [x] enable AppTP
- [x] go to App tracking protection scree and click on "report an issue"
- [x] select any app and report
- [x] verify the breakage metadata contains the field `reportedAppUnprotectedByDefault` inside the `vpnTrackerLists` object
- [x] verify the field is `true` or `false` depending on whether the reported app is in the exclusion list or not