Fix TracerPacket memory leak (#1702)
Task/Issue URL: https://app.asana.com/0/0/1201701653646856/f

### Description
This PR fixes the memory leak caused by tracer buffer.
Tracer packets have an associated `ByteBuffer` needed to inser the packet in the network-to-device queue. One we take the packet out and process it, we were not releasing it. Causing a memory leak.

This PR:
* releases the bytebuffer associated to the tracer packet
* Incidentally
  * adds some log statements for RUN write errors
  * double ensures that `acquire` buffers are cleared so that they're ready to be used

### Steps to test this PR
Smoke tests for AppTP