solve race condition in vpn state query (#1783)
Task/Issue URL: https://app.asana.com/0/1174433894299346/1201913071568577

Description
In AppTP: Improve communication for AppTP disables we changed how the Trackers screen toggle receives its information.

Before we polled the VpnService state every second, now we relay on a callback that tells us when the VPN as started / stopped.

After doing the change, we've seen cases where the toggle and the vpn state are out of sync. The toggle shows AppTP as disabled, but the VPN service is running.