Exclude VPN routes to fix wi-fi calling (#1787)
Task/Issue URL: https://app.asana.com/0/1199371158290272/1201438077555953/f

### Description
There seems to be a bug in some OEMs that breaks wi-fi calling when there's an active android VPN.

Samsung seems to be one of the OEMs affected. As Samsung footprint is very high amongst our users, we fix this issue by adding exclusion routes to the t-mobile and verizon ePDG servers.

### Steps to test this PR
Perform smoke tests for AppTP

Verify that the `VpnRoutes` contain the following routes:
* T-mobile wi-fi calling
 * 66.94.2.0/24  -> 66.94.2.255
 * 66.94.6.0/23  -> 66.94.7.255
 * 66.94.8.0/22  -> 66.94.11.255 
 * 208.54.0.0/16 -> 208.54.255.255
* Verizon wi-fi calling
 * 66.174.0.0/16 -> 66.174.255.255
 * 69.82.0.0/15  -> 69.83.255.255
 * 69.96.0.0/13  -> 69.103.255.255
 * 70.192.0.0/11 -> 70.223.255.255
 * 72.96.0.0/9   -> 72.127.255.255
 * 75.192.0.0/9  -> 75.255.255.255
 * 97.0.0.0/10   -> 97.63.255.255
 * 97.128.0.0/9  -> 97.255.255.255
 * 174.192.0.0/9 -> 174.255.255.255