Dedup byte buffer pool release mechanism (#1806)
Task/Issue URL: https://app.asana.com/0/1199371158290272/1201949747150726/f

### Description
Releasing the byte buffers that back the packets is a tedious process, and it can happen that we try to release the same buffer twice.

As the buffers a direct `ByteBuffers`, when released, they end up in a pool of available buffers, which is implemented by a linked queue. It can happen that the same buffer is released twice, resulting in a duplicated entry in the queue, eventually ending up in possible buffer corruption as it can be allocated twice.

This PR ensures that it won't be possible to release the same buffer twice.

### Steps to test this PR
Smoke tests in AppTP.

The newly added tests pass.