Detect and set private DNS in AppTP (#1796)
Task/Issue URL: https://app.asana.com/0/1199371158290272/1201931173358074/f

### Description
Try to fix the issues with private DNS settings.

When users go to Android settings and set the private DNS, the systems expects to use DoT[H] however the VPN default DNS will not support it. And that will create issues in certain cases.

The PR listens for the `private_dns_mode` setting and adds those as well to the VPN configuration.

This is done for now in internal builds only to give it some bake time


### Steps to test this PR

_Check private DNS is added to VPN config in internal builds_
- [x] install from this branch the internal flavor
- [x] Enable AppTP and filter logcat by `Setting private DNS`
- [x] Go to Android setting `Private DNS`, configure it to `manual` and set it to `one.one.one.one`
- [x] Click `save`
- [x] Verify the `Setting private DNS:...` log appears in the console
- [x] Go to diagnostics screen and verify the DNS(es) also appear in there
- [x] Repeate this test using the play flavor
- [x] verify the `Setting private DNS:...` log does not appear and that in diagnostics no additional DNS appears

Repeat this tests above for both WIFI and CELLULAR connections. The one difference may be that in CELLULAR you may also see IPv6 DNS addresses.