Add delay before showing autofill save/update prompts (#2073)
Without this delay, the form submission / navigation web events can be blocked, meaning we save the credentials but the user is left on the form, needing to hit the submission button again.

<!--
Note: This checklist is a reminder of our shared engineering expectations. 
The items in Bold are required
If your PR involves UI changes:
    1. Upload screenshots or screencasts that illustrate the changes before / after
    2. Add them under the UI changes section (feel free to add more columns if needed)
    3. Make sure these changes are tested in API 23 and API 26
If your PR does not involve UI changes, you can remove the **UI changes** section
-->

Task/Issue URL: https://app.asana.com/0/488551667048375/1202612639205162/f

### Description
As noted in the linked task, we can hit a problem with forms not submitting correctly if we immediately show a DialogFragment from the @Javainterface function. This is a strange problem, but also one hit on other platforms like iOS.

The workaround is to introduce a short delay before showing the DialogFragments for saving or updating a credential

Also included is a small tidy-up of coroutine usage around these areas, which weren't quite right before.

### Steps to test this PR

1. Visit https://trello.com/login
2. Enter credentials and hit login button
3. Verify save dialog is shown
4. Also verify the form was submitted (either you'll be logged in, or you'll at least see error message about incorrect login). On `develop`, the page wouldn't change and you'd have to tap login button again