Add DuckDuckgoFragment and use it (#2137)
Task/Issue URL: https://app.asana.com/0/488551667048375/1202717423070878/f

### Description
`Fragments` currently have an associated dagger component but are not dagger injectors.

Similar to what we did for Activities, create a `DuckDuckGoFragment` and `DaggerFragment` that contains all the logic so that we don't need to repeat it in every single fragment.

### Steps to test this PR
Smoke tests for:
* autofill fragments
* email fragments
* AppTP fragments
* browser fragment
* feedback fragment