Extract configuration related logic out from BrowserAutofill (#2259)
Task/Issue URL: https://app.asana.com/0/488551667048375/1202922320754385/f

Description

Prior to this change, InlineBrowserAutofill does 2 things: 1.) configures the webview for the specific site being loaded and 2.) handles injection of credentials. Because of this 2 purposes, BrowserAutofill is injected in 2 places creating 2 instances for every tab. The problem with this is that AutofillStoredBackJavascriptInterface together with majority of its dependencies are duplicated when we only need them for purpose 2.

To solve it in this PR, we separate the 2 purposes into 2 separate group of classes. Thus we only inject exactly what we need in the BrowserWebViewClient.

Steps to test this PR

  Smoke test Autofill
  Smoke test Email