Autofill defaults to off generally, but overridden to on for internal testers (#2520)
<!--
Note: This checklist is a reminder of our shared engineering
expectations.
The items in Bold are required
If your PR involves UI changes:
1. Upload screenshots or screencasts that illustrate the changes before
/ after
2. Add them under the UI changes section (feel free to add more columns
if needed)
    3. Make sure these changes are tested in API 23 and API 26
If your PR does not involve UI changes, you can remove the **UI
changes** section
-->

Task/Issue URL:
https://app.asana.com/0/488551667048375/1203328858503338/f

### Description

Autofill is currently gated behind two mechanisms
- global on/off switch
- available to internal testing only

The current default of the switch is `enabled`; this PR changes that to
be enabled if an internal tester, false otherwise. The internal testing
gate still applies, so really there is no behaviour change, but it's
preparation for when the internal tester gate is removed.