Improve autofill subdomain/url handling (#2552)
<!--
Note: This checklist is a reminder of our shared engineering
expectations.
The items in Bold are required
If your PR involves UI changes:
1. Upload screenshots or screencasts that illustrate the changes before
/ after
2. Add them under the UI changes section (feel free to add more columns
if needed)
    3. Make sure these changes are tested in API 23 and API 26
If your PR does not involve UI changes, you can remove the **UI
changes** section
-->

Task/Issue URL:
https://app.asana.com/0/488551667048375/1203418026431060/f

### Description
Updates the autofill URL matching logic to the linked task. Summarizing:

- better use of e-tld+1 for comparisons
- better handling of subdomains

To achieve this fully, we now filter for matching credentials in two
steps:

- at the DB level first using etld+1 
- in Kotlin, where we further filter the results based on subdomains

### Steps to test this PR

Each of the scenarios listed in the spreadsheet should be tested:
[Autofill domain matching
rules](https://app.asana.com/0/1203115518054842/1203386153863287/f)

- For each row, make sure you have a saved credential in the form
listed. Then visit the site detailed for it and validate if autofill is
offered, and whether that matches the expected state (✅ meaning we
should autofill, ❌ meaning we should not)