Reset enabledTime on manual start (#38)
<!--
Note: This checklist is a reminder of our shared engineering
expectations.
The items in Bold are required
If your PR involves UI changes:
1. Upload screenshots or screencasts that illustrate the changes before
/ after
2. Add them under the UI changes section (feel free to add more columns
if needed)
    3. Make sure these changes are tested in API 23 and API 26
If your PR does not involve UI changes, you can remove the **UI
changes** section
-->

Task/Issue URL: https://app.asana.com/0/0/1203926831020616/f

### Description
This PR ensures that when NetP is manually started that we reset
enabledTimeInMillis. Previously we are relying on VpnStopReason when the
value is SELF_STOP to detect that the service has been manually disabled
the user. However, this reason is not emitted when user disables NetP
and has AppTP still enabled since it emits RESTART instead.

### Steps to test this PR

_Only Netp_
- [x] Install this build
- [x] Enabled NetP
- [x] Verify that NetP is enabled and timing is running. Wait for 10
seconds.
- [x] Disable NetP
- [x] Re-enable NetP  again.
- [x] Verify that timer starts from 00:00:00

_Both AppTP and Netp_
- [x] Install this build
- [x] Enable AppTP
- [x] Enable NetP
- [x] Verify that NetP is enabled and timing is running. Wait for 10
seconds.
- [x] Disable NetP
- [x] Re-enable NetP  again.
- [x] Verify that timer starts from 00:00:00
- [x] Disable AppTP
- [x] Verify that timer in Netp is still running and is not reset.
- [x] Re-enable AppTP.
- [x] Verify that timer in Netp is still running and is not reset.