DAU Pixel (#2999)
<!--
Note: This checklist is a reminder of our shared engineering
expectations.
The items in Bold are required
If your PR involves UI changes:
1. Upload screenshots or screencasts that illustrate the changes before
/ after
2. Add them under the UI changes section (feel free to add more columns
if needed)
    3. Make sure these changes are tested in API 23 and API 26
If your PR does not involve UI changes, you can remove the **UI
changes** section
-->

Task/Issue URL:
https://app.asana.com/0/1157893581871903/1204238605019198

### Description
This PR adds a new DAU pixel for browser specific features. The pixel is
sent once a day with the enabled / disabled state of features. Features
can use the `BrowserFeatureStateReporterPlugin` to report state to the
pixel.

### Steps to test this PR

_Fresh install_
- [x] Launch app
- [x] Verify `m_browser_feature_daily_active_user_d` is sent with
`default_browser=0 `as parameter
- [x] Set browser as default during Dax Onboarding
- [x] Verify `m_browser_feature_daily_active_user_d` is sent with
`default_browser=1 `as parameter
- [ ] close and re-launch the app
- [ ] Verify `m_browser_feature_daily_active_user_d` is NOT sent
- [ ] Set a different browser as default browser
- [ ] close and re-launch the app
- [ ] Verify `m_browser_feature_daily_active_user_d` is NOT sent